requirements.txt
streamlit
google-generativeai
openai>=1.0.0
python-dotenv
SpeechRecognition